backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==5.1.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
facturama==2.0.5.1
fonttools==4.55.6
gunicorn==23.0.0
idna==3.10
nose==1.3.7
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyphen==0.17.2
pytz==2024.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.8.2
zopfli==0.2.3.post1
